fixes #6799 so that we reuse the current namespace OOTB for arquillian tests and can read fabric8-environments for mapping a logical environment name to a physical namespace

+++ components/fabric8-arquillian/src/main/java/io/fabric8/arquillian/utils/Util.java
@@ -101,10 +101,12 @@ public class Util {
             waitUntilWeCanDestroyNamespace(session);
             List<Throwable> errors = new ArrayList<>();
             cleanupAllMatching(client, session, errors);
-            try {
-                controller.deleteNamespace(session.getNamespace());
-            } catch (Exception e) {
-                errors.add(e);
+            if (configuration.isCreateNamespaceForTest()) {
+                try {
+                    controller.deleteNamespace(session.getNamespace());
+                } catch (Exception e) {
+                    errors.add(e);
+                }
             }
             List<Throwable> exceptions = new ArrayList<>();
             for (Throwable exception : errors) {
